Split utils.js and add basic filtered download capability
