+++ src/download.js
@@ -1,4 +1,5 @@
 import JSZip from 'jszip';
+import {coordinatesToBbox, isBboxInsideAnotherOne, areBboxIntersectings, filterCsvString} from './utils.js';
 
 
 /**
@@ -21,36 +22,6 @@ import JSZip from 'jszip';
  * @property {number[]} extent
  */
 
-/**
- * @param {number[]} extent [xmin, ymin, xmax, ymax]
- * @param {number} x
- * @param {number} y
- * @return {boolean} whether [x, y] is inside the extent
- */
-export function containsXY(extent, x, y) {
-  return extent[0] <= x && x <= extent[2] && extent[1] <= y && y <= extent[3];
-}
-
-/**
- *
- * @param {string} str
- * @param {number[]} bbox4326
- * @return {string}
- */
-export function filterCsvString(str, bbox4326) {
-  const lines = str.split('\n').filter((l, idx) => {
-    if (idx === 0) {
-      return true;
-    }
-    const columns = l.split(',');
-    const x4326 = +columns[3].trim();
-    const y4326 = +columns[4].trim();
-    const inside = containsXY(bbox4326, x4326, y4326);
-    return inside;
-  });
-  const filteredString = lines.join('\n');
-  return filteredString;
-}
 
 /**
  * Create a ZIP containing values like:
@@ -60,9 +31,9 @@ export function filterCsvString(str, bbox4326) {
  */
 export function createZipFromData(pieces) {
   const zip = new JSZip();
-  zip.folder('swissgeol-data');
+  const subZip = zip.folder('swissgeol-data');
   for (const {layer, filename, content} of pieces) {
-    zip.folder(layer).file(filename, content);
+    subZip.folder(layer).file(filename, content);
   }
   return zip;
 }
@@ -87,20 +58,6 @@ async function handleCSV(spec, bbox, fetcher) {
   };
 }
 
-/**
- *
- * @param {number[]} extent1
- * @param {number[]} extent2
- * @return {boolean}
- */
-export function isBboxIntersecting(extent1, extent2) {
-  // From OL6
-  return (
-    extent1[0] <= extent2[2] &&
-    extent1[2] >= extent2[0] &&
-    extent1[1] <= extent2[3] &&
-    extent1[3] >= extent2[1]
-  );}
 
 /**
  *
@@ -117,22 +74,7 @@ export async function getIndex(indices, spec, fetcher) {
       .then(geojson => geojson.features.map(f => {
         const filename = f.properties.filename;
         // [xmin, ymin, xmax, ymax]
-        const e = [Infinity, Infinity, -Infinity, -Infinity];
-        for (const c of f.geometry.coordinates[0]) {
-          if (c[0] < e[0]) {
-            e[0] = c[0];
-          }
-          if (c[1] < e[1]) {
-            e[1] = c[1];
-          }
-          if (c[0] > e[2]) {
-            e[2] = c[0];
-          }
-          if (c[1] > e[3]) {
-            e[3] = c[1];
-          }
-        }
-        const extent = e;
+        const extent = coordinatesToBbox(f.geometry.coordinates[0]);
         return {
           filename,
           extent
@@ -153,8 +95,8 @@ export async function getIndex(indices, spec, fetcher) {
 export async function* createIndexedDataGenerator(indices, spec, bbox, fetcher) {
   const index = await getIndex(indices, spec, fetcher);
   const path = spec.url.substr(0, spec.url.lastIndexOf('/'));
-  for (const {filename, extent} of index) {
-    if (isBboxIntersecting(extent, bbox)) {
+  for await (const {filename, extent} of index) {
+    if (areBboxIntersectings(extent, bbox)) {
       const content = await fetcher(path + '/' + filename).then(r => r.arrayBuffer());
       yield {
         layer: spec.layer,
@@ -174,7 +116,7 @@ export async function* createIndexedDataGenerator(indices, spec, bbox, fetcher) 
  */
 export async function* createDataGenerator(specs, bbox, fetcher = fetch) {
   const indices = {};
-  for (const spec of specs) {
+  for await (const spec of specs) {
     switch (spec.type) {
       case 'csv':
         yield await handleCSV(spec, bbox, fetcher);
